接口调试

+++ src/main/java/com/sulongfei/jump/dto/RoomDTO.java
@@ -41,9 +41,9 @@ public class RoomDTO extends BaseDTO {
 
     private Integer ticketNum;
 
-    private BigDecimal prizeProbability;
+    private Integer prizeProbability;
 
-    private BigDecimal premiumProportion;
+    private Integer premiumProportion;
 
     private String picture1;
